# apps/docs-site/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Claustrum Memory Core API",
    "version": "0.1.3",
    "description": "Auto-generated from memory-core route declarations and inline Zod validations."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    }
  },
  "paths": {
...
    "/v1/monorepo-subproject-policies": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "GET /v1/monorepo-subproject-policies",
        "operationId": "get_v1_monorepo_subproject_policies",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "workspace_key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repo_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-settings-mapping-routes.ts"
      },
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "POST /v1/monorepo-subproject-policies",
        "operationId": "post_v1_monorepo_subproject_policies",
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-source-file": "apps/memory-core/src/http/routes/core-settings-mapping-routes.ts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workspace_key": {
                    "type": "string"
                  },
                  "repo_key": {
                    "type": "string"
                  },
                  "subpath": {
                    "type": "string"
                  },
                  "enabled": {
                    "type": "boolean"
                  },
                  "reason": {
                    "type": "string"
                  }
                },
                "required": [
                  "workspace_key",
                  "repo_key",
                  "subpath"
                ]
              }
            }
          }
        }
      }
    },
    "/v1/monorepo-subproject-policies/{id}": {
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "PATCH /v1/monorepo-subproject-policies/:id",
        "operationId": "patch_v1_monorepo_subproject_policies_id",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-settings-mapping-routes.ts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workspace_key": {
                    "type": "string"
                  },
                  "enabled": {
                    "type": "boolean"
                  },
                  "reason": {
                    "type": "string"
                  }
                },
                "required": [
                  "workspace_key",
                  "enabled"
                ]
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Settings"
        ],
        "summary": "DELETE /v1/monorepo-subproject-policies/:id",
        "operationId": "delete_v1_monorepo_subproject_policies_id",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "workspace_key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-settings-mapping-routes.ts"
      }
    },
...
  },
...
}